backend/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
torch
transformers
rank-bm25
groq
python-dotenv
